nvmap: Fix type casting issue
CERT-C is reporting following issue:
- Casting nvmap_dev->pool.count + nvmap_dev->pool.to_zero from unsigned
int to int without checking its value may result in lost or
misinterpreted data.
Fix this by changing the data type of total to unsigned long.

CID 456842
Bug 3745813

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2816556
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/video/tegra/nvmap/nvmap_pp.c
@@ -478,7 +478,7 @@ u32 nvmap_page_pool_fill_lots(struct nvmap_page_pool *pool,
 
 ulong nvmap_page_pool_get_unused_pages(void)
 {
-	int total = 0;
+	unsigned long total = 0;
 
 	if (!nvmap_dev)
 		return 0;
